requirements.txt
parsel
prettytable
psycopg2
PyYAML
requests
SQLAlchemy
SQLAlchemy-Utils
tornado
